Move exceptions to new package
