+++ vlog4j-core/src/main/java/org/semanticweb/vlog4j/core/exceptions/ReasonerStateException.java
@@ -18,7 +18,7 @@
  * #L%
  */
 
-package org.semanticweb.vlog4j.core.reasoner.exceptions;
+package org.semanticweb.vlog4j.core.exceptions;
 
 import java.text.MessageFormat;
