+++ vlog4j-core/src/main/java/org/semanticweb/vlog4j/core/exceptions/VLog4jException.java
@@ -1,4 +1,4 @@
-package org.semanticweb.vlog4j.core.reasoner.exceptions;
+package org.semanticweb.vlog4j.core.exceptions;
 
 /*-
  * #%L
