+++ vlog4j-examples/src/main/java/org/semanticweb/vlog4j/examples/SimpleReasoningExample.java
@@ -22,8 +22,8 @@ package org.semanticweb.vlog4j.examples;
 
 import java.io.IOException;
 
+import org.semanticweb.vlog4j.core.exceptions.VLog4jException;
 import org.semanticweb.vlog4j.core.reasoner.Reasoner;
-import org.semanticweb.vlog4j.core.reasoner.exceptions.VLog4jException;
 import org.semanticweb.vlog4j.parser.ParsingException;
 import org.semanticweb.vlog4j.parser.RuleParser;
