+++ vlog4j-parser/src/main/java/org/semanticweb/vlog4j/parser/RuleParser.java
@@ -24,9 +24,9 @@ import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.util.List;
 
+import org.semanticweb.vlog4j.core.exceptions.PrefixDeclarationException;
 import org.semanticweb.vlog4j.core.model.api.Literal;
 import org.semanticweb.vlog4j.core.model.api.PositiveLiteral;
-import org.semanticweb.vlog4j.core.model.api.PrefixDeclarationException;
 import org.semanticweb.vlog4j.core.model.api.Rule;
 import org.semanticweb.vlog4j.parser.RuleParserBase.FormulaContext;
 import org.semanticweb.vlog4j.parser.javacc.JavaCCRuleParser;
